+++ src/frago9876543210/forms/bootstrap/Main.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace forms\bootstrap;
+namespace frago9876543210\forms\bootstrap;
 
 use pocketmine\plugin\PluginBase;
