+++ src/frago9876543210/forms/element/StepSlider.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace forms\element;
+namespace frago9876543210\forms\element;
 
 /** @phpstan-extends BaseSelector<int> */
-class Dropdown extends BaseSelector{
+class StepSlider extends BaseSelector{
 
-	protected function getType() : string{ return "dropdown"; }
+	protected function getType() : string{ return "step_slider"; }
 
 	protected function serializeElementData() : array{
 		return [
-			"options" => $this->options,
+			"steps" => $this->options,
 			"default" => $this->default,
 		];
 	}
