+++ src/frago9876543210/forms/element/Toggle.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace forms\element;
+namespace frago9876543210\forms\element;
 
 use pocketmine\form\FormValidationException;
 use function gettype;
